# Various performance improvements.