--- a/.changes/next-release/feature-AWSSDKforJavav2-49d3cd6.json
+++ b/.changes/next-release/feature-AWSSDKforJavav2-49d3cd6.json
@@ -1,0 +1,5 @@
+{
+    "type": "feature",
+    "category": "AWS SDK for Java v2",
+    "description": "Various performance improvements."
+}