--- a/core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/http/pipeline/stages/utils/RetryableStageHelper.java
+++ b/core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/http/pipeline/stages/utils/RetryableStageHelper.java
@@ -138,13 +138,11 @@ public class RetryableStageHelper {
         Integer availableRetryCapacity = TokenBucketRetryCondition.getCapacityForExecution(context.executionAttributes())
                                                                   .map(TokenBucketRetryCondition.Capacity::capacityRemaining)
                                                                   .orElse(null);
-
+        String headerValue = (attemptNumber - 1) + "/" +
+                             lastBackoffDelay.toMillis() + "/" +
+                             (availableRetryCapacity != null ? availableRetryCapacity : "");
         return request.toBuilder()
-                      .putHeader(SDK_RETRY_INFO_HEADER,
-                                 String.format("%s/%s/%s",
-                                               attemptNumber - 1,
-                                               lastBackoffDelay.toMillis(),
-                                               availableRetryCapacity != null ? availableRetryCapacity : ""))
+                      .putHeader(SDK_RETRY_INFO_HEADER, headerValue)
                       .build();
     }