… http-client-spi/src/test/java/software/amazon/awssdk/http/SdkHttpRequestResponseTest.java
… utils/src/main/java/software/amazon/awssdk/utils/CollectionUtils.java
… utils/src/main/java/software/amazon/awssdk/utils/StringUtils.java
… utils/src/main/java/software/amazon/awssdk/utils/UnmodifiableMapOfLists.java
… utils/src/main/java/software/amazon/awssdk/utils/http/SdkHttpUtils.java
… utils/src/test/java/software/amazon/awssdk/utils/CollectionUtilsTest.java